requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
bullet-safety-gym==1.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
cloudpickle==3.1.0
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.0a10
-e git+https://gitclone.com/github.com/Farama-Foundation/D4RL.git@71a9549f2091accff93eeff68f1f3ab2c0e0a288#egg=D4RL
decorator==5.1.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.26
docker-pycreds==0.4.0
dsrl==0.1.0
etils==1.5.2
eval_type_backport==0.2.2
Farama-Notifications==0.0.4
fasteners==0.19
flax==0.8.5
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.8.0
grpcio==1.69.0
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.12.1
humanize==4.11.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jax-jumpy==1.0.0
jaxlib==0.4.30
kiwisolver==1.4.7
labmaze==1.0.6
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mjrl @ git+https://gitclone.com/github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
ml_collections==0.1.1
ml_dtypes==0.5.0
msgpack==1.1.0
mujoco==3.2.7
mujoco-py==2.1.2.14
nest-asyncio==1.6.0
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.6.0.74
nvidia-cufft-cu12==11.3.0.4
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.6.85
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.6.4
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
pybullet==3.2.6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.1.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
safety-gymnasium==0.4.0
scipy==1.13.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-probability==0.21.0
tensorstore==0.1.69
termcolor==2.5.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.1